+++ binary-decode.go
@@ -17,7 +17,7 @@ import (
 // function calls decodeReflectBinary*, and generally those functions should
 // only call this one, for the prefix bytes are consumed here when present.
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Value, fopts FieldOptions, bare bool) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -25,60 +25,10 @@ func (cdc *Codec) decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Valu
 		panic("should not happen")
 	}
 	if printLog {
-		spew.Printf("(D) decodeReflectBinary(bz: %X, info: %v, rv: %#v (%v), opts: %v)\n",
-			bz, info, rv.Interface(), rv.Type(), opts)
+		spew.Printf("(D) decodeReflectBinary(bz: %X, info: %v, rv: %#v (%v), fopts: %v)\n",
+			bz, info, rv.Interface(), rv.Type(), fopts)
 		defer func() {
 			fmt.Printf("(D) -> n: %v, err: %v\n", n, err)
-		}()
-	}
-
-	// TODO Read the disamb bytes here if necessary.
-	// e.g. rv isn't an interface, and
-	// info.ConcreteType.AlwaysDisambiguate.  But we don't support
-	// this yet.
-
-	// Read prefix+typ3 bytes if registered.
-	if info.Registered {
-		if len(bz) < PrefixBytesLen {
-			err = errors.New("EOF skipping prefix bytes.")
-			return
-		}
-		// Check prefix bytes.
-		prefix3 := NewPrefixBytes(bz[:PrefixBytesLen])
-		var prefix, typ = prefix3.SplitTyp3()
-		if info.Prefix != prefix {
-			panic("should not happen")
-		}
-		// Check that typ3 in prefix bytes is correct.
-		err = checkTyp3(info.Type, typ, opts)
-		if err != nil {
-			return
-		}
-		// Consume prefix.  Yum.
-		bz = bz[PrefixBytesLen:]
-		n += PrefixBytesLen
-	}
-
-	_n := 0
-	_n, err = cdc._decodeReflectBinary(bz, info, rv, opts)
-	slide(&bz, &n, _n)
-	return
-}
-
-// CONTRACT: any immediate disamb/prefix bytes have been consumed/stripped.
-// CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
-	if !rv.CanAddr() {
-		panic("rv not addressable")
-	}
-	if info.Type.Kind() == reflect.Interface && rv.Kind() == reflect.Ptr {
-		panic("should not happen")
-	}
-	if printLog {
-		spew.Printf("(_) _decodeReflectBinary(bz: %X, info: %v, rv: %#v (%v), opts: %v)\n",
-			bz, info, rv.Interface(), rv.Type(), opts)
-		defer func() {
-			fmt.Printf("(_) -> n: %v, err: %v\n", n, err)
 		}()
 	}
 	var _n int
@@ -103,7 +53,7 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 		if err != nil {
 			return
 		}
-		_n, err = cdc._decodeReflectBinary(bz, rinfo, rrv, opts)
+		_n, err = cdc.decodeReflectBinary(bz, rinfo, rrv, fopts, bare)
 		if slide(&bz, &n, _n) && err != nil {
 			return
 		}
@@ -123,17 +73,17 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 	// Complex
 
 	case reflect.Interface:
-		_n, err = cdc.decodeReflectBinaryInterface(bz, info, rv, opts)
+		_n, err = cdc.decodeReflectBinaryInterface(bz, info, rv, fopts, bare)
 		n += _n
 		return
 
 	case reflect.Array:
 		ert := info.Type.Elem()
 		if ert.Kind() == reflect.Uint8 {
-			_n, err = cdc.decodeReflectBinaryByteArray(bz, info, rv, opts)
+			_n, err = cdc.decodeReflectBinaryByteArray(bz, info, rv, fopts)
 			n += _n
 		} else {
-			_n, err = cdc.decodeReflectBinaryArray(bz, info, rv, opts)
+			_n, err = cdc.decodeReflectBinaryArray(bz, info, rv, fopts, bare)
 			n += _n
 		}
 		return
@@ -141,16 +91,16 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 	case reflect.Slice:
 		ert := info.Type.Elem()
 		if ert.Kind() == reflect.Uint8 {
-			_n, err = cdc.decodeReflectBinaryByteSlice(bz, info, rv, opts)
+			_n, err = cdc.decodeReflectBinaryByteSlice(bz, info, rv, fopts)
 			n += _n
 		} else {
-			_n, err = cdc.decodeReflectBinarySlice(bz, info, rv, opts)
+			_n, err = cdc.decodeReflectBinarySlice(bz, info, rv, fopts, bare)
 			n += _n
 		}
 		return
 
 	case reflect.Struct:
-		_n, err = cdc.decodeReflectBinaryStruct(bz, info, rv, opts)
+		_n, err = cdc.decodeReflectBinaryStruct(bz, info, rv, fopts, bare)
 		n += _n
 		return
 
@@ -159,7 +109,7 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 
 	case reflect.Int64:
 		var num int64
-		if opts.BinVarint {
+		if fopts.BinVarint {
 			num, _n, err = DecodeVarint(bz)
 			if slide(&bz, &n, _n) && err != nil {
 				return
@@ -215,7 +165,7 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 
 	case reflect.Uint64:
 		var num uint64
-		if opts.BinVarint {
+		if fopts.BinVarint {
 			num, _n, err = DecodeUvarint(bz)
 			if slide(&bz, &n, _n) && err != nil {
 				return
@@ -280,7 +230,7 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 
 	case reflect.Float64:
 		var f float64
-		if !opts.Unsafe {
+		if !fopts.Unsafe {
 			err = errors.New("Float support requires `amino:\"unsafe\"`.")
 			return
 		}
@@ -293,7 +243,7 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 
 	case reflect.Float32:
 		var f float32
-		if !opts.Unsafe {
+		if !fopts.Unsafe {
 			err = errors.New("Float support requires `amino:\"unsafe\"`.")
 			return
 		}
@@ -320,7 +270,7 @@ func (cdc *Codec) _decodeReflectBinary(bz []byte, info *TypeInfo, rv reflect.Val
 }
 
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinaryInterface(bz []byte, iinfo *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinaryInterface(bz []byte, iinfo *TypeInfo, rv reflect.Value, fopts FieldOptions, bare bool) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -338,15 +288,21 @@ func (cdc *Codec) decodeReflectBinaryInterface(bz []byte, iinfo *TypeInfo, rv re
 		return
 	}
 
-	// Consume disambiguation / prefix+typ3 bytes.
-	disamb, hasDisamb, prefix, typ, hasPrefix, isNil, _n, err := DecodeDisambPrefixBytes(bz)
-	if slide(&bz, &n, _n) && err != nil {
-		return
+	if !bare {
+		// Read byte-length prefixed byteslice.
+		var buf, _n = []byte(nil), int(0)
+		buf, _n, err = DecodeByteSlice(bz)
+		if slide(&bz, nil, _n) && err != nil {
+			return
+		}
+		// This is a trick for debuggability -- we slide on &n more later.
+		n += UvarintSize(uint64(len(buf)))
+		bz = buf
 	}
 
-	// Special case for nil.
-	if isNil {
-		rv.Set(iinfo.ZeroValue)
+	// Consume disambiguation / prefix bytes.
+	disamb, hasDisamb, prefix, hasPrefix, _n, err := DecodeDisambPrefixBytes(bz)
+	if slide(&bz, &n, _n) && err != nil {
 		return
 	}
 
@@ -363,20 +319,20 @@ func (cdc *Codec) decodeReflectBinaryInterface(bz []byte, iinfo *TypeInfo, rv re
 		return
 	}
 
-	// Check and consume typ3 byte.
-	// It cannot be a typ4 byte because it cannot be nil.
-	err = checkTyp3(cinfo.Type, typ, opts)
-	if err != nil {
-		return
-	}
-
 	// Construct the concrete type.
 	var crv, irvSet = constructConcreteType(cinfo)
 
 	// Decode into the concrete type.
-	_n, err = cdc._decodeReflectBinary(bz, cinfo, crv, opts)
+	_n, err = cdc.decodeReflectBinary(bz, cinfo, crv, fopts, true)
 	if slide(&bz, &n, _n) && err != nil {
 		rv.Set(irvSet) // Helps with debugging
+		return
+	}
+
+	// Earlier, we set bz to the byteslice read from buf.
+	// Ensure that all of bz was consumed.
+	if len(bz) > 0 {
+		err = errors.New("bytes left over after reading interface contents")
 		return
 	}
 
@@ -390,7 +346,7 @@ func (cdc *Codec) decodeReflectBinaryInterface(bz []byte, iinfo *TypeInfo, rv re
 }
 
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinaryByteArray(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinaryByteArray(bz []byte, info *TypeInfo, rv reflect.Value, fopts FieldOptions) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -427,7 +383,7 @@ func (cdc *Codec) decodeReflectBinaryByteArray(bz []byte, info *TypeInfo, rv ref
 }
 
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinaryArray(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinaryArray(bz []byte, info *TypeInfo, rv reflect.Value, fopts FieldOptions, bare bool) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -442,65 +398,106 @@ func (cdc *Codec) decodeReflectBinaryArray(bz []byte, info *TypeInfo, rv reflect
 		panic("should not happen")
 	}
 	length := info.Type.Len()
-	einfo := (*TypeInfo)(nil)
-	einfo, err = cdc.getTypeInfo_wlock(ert)
+	einfo, err := cdc.getTypeInfo_wlock(ert)
 	if err != nil {
 		return
 	}
 
-	// Check and consume typ4 byte.
-	var ptr, _n = false, int(0)
-	ptr, _n, err = decodeTyp4AndCheck(ert, bz, opts)
-	if slide(&bz, &n, _n) && err != nil {
-		return
+	if !bare {
+		// Read byte-length prefixed byteslice.
+		var buf, _n = []byte(nil), int(0)
+		buf, _n, err = DecodeByteSlice(bz)
+		if slide(&bz, nil, _n) && err != nil {
+			return
+		}
+		// This is a trick for debuggability -- we slide on &n more later.
+		n += UvarintSize(uint64(len(buf)))
+		bz = buf
 	}
 
-	// Read number of items.
-	var count = uint64(0)
-	count, _n, err = DecodeUvarint(bz)
-	if slide(&bz, &n, _n) && err != nil {
-		return
-	}
-	if int(count) != length {
-		err = fmt.Errorf("Expected num items of %v, decoded %v", length, count)
-		return
-	}
-
-	// NOTE: Unlike decodeReflectBinarySlice,
-	// there is nothing special to do for
-	// zero-length arrays.  Is that even possible?
-
-	// Read each item.
-	for i := 0; i < length; i++ {
-		var erv, _n = rv.Index(i), int(0)
-		// Maybe read nil.
-		if ptr {
-			numNil := int64(0)
-			numNil, _n, err = decodeNumNilBytes(bz)
+	// If elem is not already a ByteLength type, read in packed form.
+	// This is a Proto wart due to Proto backwards compatibility issues.
+	// Amino2 will probably migrate to use the List typ3.
+	typ3 := typeToTyp3(einfo.Type, fopts)
+	if typ3 != Typ3_ByteLength {
+		// Read elements in packed form.
+		for i := 0; i < length; i++ {
+			var erv, _n = rv.Index(i), int(0)
+			_n, err = cdc.decodeReflectBinary(bz, einfo, erv, fopts, false)
+			if slide(&bz, &n, _n) && err != nil {
+				err = fmt.Errorf("error reading array contents: %v", err)
+				return
+			}
+			// Special case when reading default value, prefer nil.
+			if erv.Kind() == reflect.Ptr {
+				_, isDefault := isDefaultValue(erv)
+				if isDefault {
+					erv.Set(reflect.Zero(erv.Type()))
+					continue
+				}
+			}
+		}
+		// Ensure that we read the whole buffer.
+		if len(bz) > 0 {
+			err = errors.New("bytes left over after reading array contents")
+			return
+		}
+	} else {
+		// Read elements in unpacked form.
+		for i := 0; i < length; i++ {
+			// Read field key (number and type).
+			var fieldNum, typ, _n = uint32(0), Typ3(0x00), int(0)
+			fieldNum, typ, _n, err = decodeFieldNumberAndTyp3(bz)
+			// Validate field number and typ3.
+			if fieldNum != fopts.BinFieldNum {
+				err = errors.New(fmt.Sprintf("expected repeated field number %v, got %v", fopts.BinFieldNum, fieldNum))
+				return
+			}
+			if typ != Typ3_ByteLength {
+				err = errors.New(fmt.Sprintf("expected repeated field type %v, got %v", Typ3_ByteLength, typ))
+				return
+			}
 			if slide(&bz, &n, _n) && err != nil {
 				return
 			}
-			if numNil == 0 {
-				// Good, continue decoding item.
-			} else if numNil == 1 {
-				// Set nil/zero.
+			// Decode the next ByteLength bytes into erv.
+			var erv = rv.Index(i)
+			// Special case if next ByteLength bytes are 0x00, set nil.
+			if len(bz) > 0 && bz[0] == 0x00 {
+				slide(&bz, &n, 1)
 				erv.Set(reflect.Zero(erv.Type()))
 				continue
-			} else {
-				panic("should not happen")
+			}
+			// Normal case, read next non-nil element from bz.
+			// In case of any inner lists in unpacked form.
+			efopts := fopts
+			efopts.BinFieldNum = 1
+			_n, err = cdc.decodeReflectBinary(bz, einfo, erv, efopts, false)
+			if slide(&bz, &n, _n) && err != nil {
+				err = fmt.Errorf("error reading array contents: %v", err)
+				return
 			}
 		}
-		// Decode non-nil value.
-		_n, err = cdc.decodeReflectBinary(bz, einfo, erv, opts)
-		if slide(&bz, &n, _n) && err != nil {
-			return
+		// Ensure that there are no more elements left,
+		// and no field number regression either.
+		// This is to provide better error messages.
+		if len(bz) > 0 {
+			var fieldNum = uint32(0)
+			fieldNum, _, _, err = decodeFieldNumberAndTyp3(bz)
+			if err != nil {
+				return
+			}
+			if fieldNum <= fopts.BinFieldNum {
+				err = fmt.Errorf("unexpected field number %v after repeated field number %v", fieldNum, fopts.BinFieldNum)
+				return
+			}
 		}
 	}
 	return
 }
 
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinaryByteSlice(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinaryByteSlice(bz []byte, info *TypeInfo, rv reflect.Value, fopts FieldOptions) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -532,7 +529,7 @@ func (cdc *Codec) decodeReflectBinaryByteSlice(bz []byte, info *TypeInfo, rv ref
 }
 
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinarySlice(bz []byte, info *TypeInfo, rv reflect.Value, opts FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinarySlice(bz []byte, info *TypeInfo, rv reflect.Value, fopts FieldOptions, bare bool) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -546,72 +543,98 @@ func (cdc *Codec) decodeReflectBinarySlice(bz []byte, info *TypeInfo, rv reflect
 	if ert.Kind() == reflect.Uint8 {
 		panic("should not happen")
 	}
-	einfo := (*TypeInfo)(nil)
-	einfo, err = cdc.getTypeInfo_wlock(ert)
+	einfo, err := cdc.getTypeInfo_wlock(ert)
 	if err != nil {
 		return
 	}
 
-	// Check and consume typ4 byte.
-	var ptr, _n = false, int(0)
-	ptr, _n, err = decodeTyp4AndCheck(ert, bz, opts)
-	if slide(&bz, &n, _n) && err != nil {
-		return
+	// Construct slice to collect decoded items to.
+	// NOTE: This is due to Proto3.  How to best optimize?
+	esrt := reflect.SliceOf(ert)
+	var srv = reflect.Zero(esrt)
+
+	if !bare {
+		// Read byte-length prefixed byteslice.
+		var buf, _n = []byte(nil), int(0)
+		buf, _n, err = DecodeByteSlice(bz)
+		if slide(&bz, nil, _n) && err != nil {
+			return
+		}
+		// This is a trick for debuggability -- we slide on &n more later.
+		n += UvarintSize(uint64(len(buf)))
+		bz = buf
 	}
 
-	// Read number of items.
-	var count = uint64(0)
-	count, _n, err = DecodeUvarint(bz)
-	if slide(&bz, &n, _n) && err != nil {
-		return
-	}
-	if int(count) < 0 {
-		err = fmt.Errorf("Impossible number of elements (%v)", count)
-		return
-	}
-	if int(count) > len(bz) { // Currently, each item takes at least 1 byte.
-		err = fmt.Errorf("Impossible number of elements (%v) compared to buffer length (%v)",
-			count, len(bz))
-		return
-	}
-
-	// Special case when length is 0.
-	// NOTE: We prefer nil slices.
-	if count == 0 {
-		rv.Set(info.ZeroValue)
-		return
-	}
-
-	// Read each item.
-	// NOTE: Unlike decodeReflectBinaryArray,
-	// we need to construct a new slice before
-	// we populate it. Arrays on the other hand
-	// reserve space in the value itself.
-	var esrt = reflect.SliceOf(ert) // TODO could be optimized.
-	var srv = reflect.MakeSlice(esrt, int(count), int(count))
-	for i := 0; i < int(count); i++ {
-		var erv, _n = srv.Index(i), int(0)
-		// Maybe read nil.
-		if ptr {
-			var numNil = int64(0)
-			numNil, _n, err = decodeNumNilBytes(bz)
+	// If elem is not already a ByteLength type, read in packed form.
+	// This is a Proto wart due to Proto backwards compatibility issues.
+	// Amino2 will probably migrate to use the List typ3.
+	typ3 := typeToTyp3(einfo.Type, fopts)
+	if typ3 != Typ3_ByteLength {
+		// Read elems in packed form.
+		for {
+			if len(bz) == 0 {
+				break
+			}
+			erv, _n := reflect.New(ert).Elem(), int(0)
+			_n, err = cdc.decodeReflectBinary(bz, einfo, erv, fopts, false)
+			if slide(&bz, &n, _n) && err != nil {
+				err = fmt.Errorf("error reading array contents: %v", err)
+				return
+			}
+			// Special case when reading default value, prefer nil.
+			if ert.Kind() == reflect.Ptr {
+				_, isDefault := isDefaultValue(erv)
+				if isDefault {
+					srv = reflect.Append(srv, reflect.Zero(ert))
+					continue
+				}
+			}
+			// Otherwise append to slice.
+			srv = reflect.Append(srv, erv)
+		}
+	} else {
+		// Read elements in unpacked form.
+		for {
+			if len(bz) == 0 {
+				break
+			}
+			// Read field key (number and type).
+			var fieldNum, typ, _n = uint32(0), Typ3(0x00), int(0)
+			fieldNum, typ, _n, err = decodeFieldNumberAndTyp3(bz)
+			// Validate field number and typ3.
+			if fieldNum < fopts.BinFieldNum {
+				err = errors.New(fmt.Sprintf("expected repeated field number %v or greater, got %v", fopts.BinFieldNum, fieldNum))
+				return
+			}
+			if fieldNum > fopts.BinFieldNum {
+				break
+			}
+			if typ != Typ3_ByteLength {
+				err = errors.New(fmt.Sprintf("expected repeated field type %v, got %v", Typ3_ByteLength, typ))
+				return
+			}
 			if slide(&bz, &n, _n) && err != nil {
 				return
 			}
-			if numNil == 0 {
-				// Good, continue decoding item.
-			} else if numNil == 1 {
-				// Set nil/zero.
+			// Decode the next ByteLength bytes into erv.
+			erv, _n := reflect.New(ert).Elem(), int(0)
+			// Special case if next ByteLength bytes are 0x00, set nil.
+			if len(bz) > 0 && bz[0] == 0x00 {
+				slide(&bz, &n, 1)
 				erv.Set(reflect.Zero(erv.Type()))
+				srv = reflect.Append(srv, erv)
 				continue
-			} else {
-				panic("should not happen")
 			}
-		}
-		// Decode non-nil value.
-		_n, err = cdc.decodeReflectBinary(bz, einfo, erv, opts)
-		if slide(&bz, &n, _n) && err != nil {
-			return
+			// Normal case, read next non-nil element from bz.
+			// In case of any inner lists in unpacked form.
+			efopts := fopts
+			efopts.BinFieldNum = 1
+			_n, err = cdc.decodeReflectBinary(bz, einfo, erv, efopts, false)
+			if slide(&bz, &n, _n) && err != nil {
+				err = fmt.Errorf("error reading array contents: %v", err)
+				return
+			}
+			srv = reflect.Append(srv, erv)
 		}
 	}
 	rv.Set(srv)
@@ -619,7 +642,7 @@ func (cdc *Codec) decodeReflectBinarySlice(bz []byte, info *TypeInfo, rv reflect
 }
 
 // CONTRACT: rv.CanAddr() is true.
-func (cdc *Codec) decodeReflectBinaryStruct(bz []byte, info *TypeInfo, rv reflect.Value, _ FieldOptions) (n int, err error) {
+func (cdc *Codec) decodeReflectBinaryStruct(bz []byte, info *TypeInfo, rv reflect.Value, _ FieldOptions, bare bool) (n int, err error) {
 	if !rv.CanAddr() {
 		panic("rv not addressable")
 	}
@@ -631,8 +654,21 @@ func (cdc *Codec) decodeReflectBinaryStruct(bz []byte, info *TypeInfo, rv reflec
 	}
 	_n := 0 // nolint: ineffassign
 
-	// The "Struct" typ3 doesn't get read here.
+  // NOTE: The "Struct" typ3 doesn't get read here.
 	// It's already implied, either by struct-key or list-element-type-byte.
+
+	if !bare {
+		// Read byte-length prefixed byteslice.
+		var buf, _n = []byte(nil), int(0)
+		buf, _n, err = DecodeByteSlice(bz)
+		if slide(&bz, nil, _n) && err != nil {
+			return
+		}
+		// This is a trick for debuggability -- we slide on &n more later.
+		n += UvarintSize(uint64(len(buf)))
+		bz = buf
+	}
+
 	switch info.Type {
 
 	case timeType:
@@ -643,10 +679,9 @@ func (cdc *Codec) decodeReflectBinaryStruct(bz []byte, info *TypeInfo, rv reflec
 			return
 		}
 		rv.Set(reflect.ValueOf(t))
-		return
 
 	default:
-		// Store the last seen field number.
+		// Track the last seen field number.
 		var lastFieldNum uint32
 		// Read each field.
 		for _, field := range info.Fields {
@@ -659,54 +694,65 @@ func (cdc *Codec) decodeReflectBinaryStruct(bz []byte, info *TypeInfo, rv reflec
 				return
 			}
 
-			// Read field key (number and type).
-			var fieldNum, typ = uint32(0), Typ3(0x00)
-			fieldNum, typ, _n, err = decodeFieldNumberAndTyp3(bz)
-			if field.BinFieldNum < fieldNum {
-				// Set nil field value.
+			// We're done if we've consumed all the bytes.
+			if len(bz) == 0 {
 				frv.Set(reflect.Zero(frv.Type()))
 				continue
-				// Do not slide, we will read it again.
 			}
-			if fieldNum == 0 {
-				// Probably a StructTerm.
-				break
-			}
-			if fieldNum <= lastFieldNum {
-				err = fmt.Errorf("encountered fieldnNum: %v, but we have already seen fieldNum: %v",
+
+			if field.UnpackedList {
+				// This is a list that was encoded unpacked, e.g.
+				// with repeated field entries for each list item.
+				_n, err = cdc.decodeReflectBinary(bz, finfo, frv, field.FieldOptions, true)
+				if slide(&bz, &n, _n) && err != nil {
+					return
+				}
+			} else {
+				// Read field key (number and type).
+				var fieldNum, typ = uint32(0), Typ3(0x00)
+				fieldNum, typ, _n, err = decodeFieldNumberAndTyp3(bz)
+				if field.BinFieldNum < fieldNum {
+					// Set zero field value.
+					frv.Set(reflect.Zero(frv.Type()))
+					continue
+					// Do not slide, we will read it again.
+				}
+        if fieldNum <= lastFieldNum {
+				  err = fmt.Errorf("encountered fieldnNum: %v, but we have already seen fieldNum: %v",
 					fieldNum, lastFieldNum)
-				return
-			}
-			lastFieldNum = fieldNum
+				  return
+			  }
+			  lastFieldNum = fieldNum
+				if slide(&bz, &n, _n) && err != nil {
+					return
+				}
 
-			if slide(&bz, &n, _n) && err != nil {
-				return
-			}
-			// NOTE: In the future, we'll support upgradeability.
-			// So in the future, this may not match,
-			// so we will need to remove this sanity check.
-			if field.BinFieldNum != fieldNum {
-				err = fmt.Errorf("expected field number %v, got %v", field.BinFieldNum, fieldNum)
-				return
-			}
-			typWanted := typeToTyp4(field.Type, field.FieldOptions).Typ3()
-			if typ != typWanted {
-				err = fmt.Errorf("expected field type %v, got %v", typWanted, typ)
-				return
-			}
+				// Validate fieldNum and typ.
+				// NOTE: In the future, we'll support upgradeability.
+				// So in the future, this may not match,
+				// so we will need to remove this sanity check.
+				if field.BinFieldNum != fieldNum {
+					err = errors.New(fmt.Sprintf("expected field number %v, got %v", field.BinFieldNum, fieldNum))
+					return
+				}
+				typWanted := typeToTyp3(finfo.Type, field.FieldOptions)
+				if typ != typWanted {
+					err = errors.New(fmt.Sprintf("expected field type %v, got %v", typWanted, typ))
+					return
+				}
 
-			// Decode field into frv.
-			_n, err = cdc.decodeReflectBinary(bz, finfo, frv, field.FieldOptions)
-			if slide(&bz, &n, _n) && err != nil {
-				return
+				// Decode field into frv.
+				_n, err = cdc.decodeReflectBinary(bz, finfo, frv, field.FieldOptions, false)
+				if slide(&bz, &n, _n) && err != nil {
+					return
+				}
 			}
 		}
 
 		var _n, fnum = 0, uint32(0)
 		var typ3 Typ3
-		// we have read all, or do we? consume the rest
+		// Consume any remaining fields.
 		for {
-
 			fnum, typ3, _n, err = decodeFieldNumberAndTyp3(bz)
 			if typ3 == Typ3_StructTerm {
 				break
@@ -726,15 +772,15 @@ func (cdc *Codec) decodeReflectBinaryStruct(bz []byte, info *TypeInfo, rv reflec
 				return
 			}
 		}
-		if slide(&bz, &n, _n) && err != nil {
-			return
-		}
 	}
 	return
 }
 
+//----------------------------------------
+// consume* for skipping struct fields
+
 // Read everything without doing anything with it. Report errors if they occur.
-func consumeAny(typ3 Typ3, bz []byte) (term bool, n int, err error) {
+func consumeAny(typ3 Typ3, bz []byte) (n int, err error) {
 	var _n int
 	switch typ3 {
 	case Typ3_Varint:
@@ -743,22 +789,11 @@ func consumeAny(typ3 Typ3, bz []byte) (term bool, n int, err error) {
 		_, _n, err = DecodeInt64(bz)
 	case Typ3_ByteLength:
 		_, _n, err = DecodeUvarint(bz)
-	case Typ3_Struct:
-		n, err = consumeStruct(bz)
-	case Typ3_StructTerm:
-		// don't slide, we'll read this later
-		term = true
 	case Typ3_4Byte:
 		_, _n, err = DecodeInt32(bz)
-	case Typ3_List:
-		_n, err = consumeList(bz)
-	case Typ3_Interface:
-		_n, err = consumeInterface(bz)
 	default:
 		panic("should not happen")
-
 	}
-
 	if err != nil {
 		// do not slide
 		return
@@ -775,73 +810,7 @@ func consumeStruct(bz []byte) (n int, err error) {
 			return
 		}
 		var term bool
-		term, _n, err = consumeAny(typ, bz)
-		if slide(&bz, &n, _n) && err != nil {
-			return
-		}
-		if term {
-			return
-		}
-	}
-	return
-}
-
-func consumeInterface(bz []byte) (n int, err error) {
-	_, _, _, typ, _, _, _n, err := DecodeDisambPrefixBytes(bz)
-	if slide(&bz, &n, _n) && err != nil {
-		return
-	}
-	_, _n, err = consumeAny(typ, bz)
-	if slide(&bz, &n, _n) && err != nil {
-		return
-	}
-	return
-}
-
-func consumeList(bz []byte) (n int, err error) {
-	// Read element Typ4.
-	if len(bz) < 1 {
-		err = errors.New("EOF reading list element typ4.")
-		return
-	}
-	var typ = Typ4(bz[0])
-	if typ&0xF0 > 0 {
-		err = errors.New("invalid list element typ4 byte")
-	}
-	if slide(&bz, &n, 1) && err != nil {
-		return
-	}
-	// Read number of elements.
-	var num, _n = uint64(0), int(0)
-	num, _n = binary.Uvarint(bz)
-	if _n < 0 {
-		_n = 0
-		err = errors.New("error decoding list length (uvarint)")
-	}
-	if slide(&bz, &n, _n) && err != nil {
-		return
-	}
-	// Read elements.
-	for i := 0; i < int(num); i++ {
-		// Maybe read nil byte.
-		if typ&0x08 != 0 {
-			if len(bz) == 0 {
-				err = errors.New("EOF reading list nil byte")
-				return
-			}
-			var nb = bz[0]
-			slide(&bz, &n, 1)
-			switch nb {
-			case 0x00:
-			case 0x01:
-				continue
-			default:
-				err = fmt.Errorf("unexpected nil pointer byte %X", nb)
-				return
-			}
-		}
-		// Read element.
-		_, _n, err = consumeAny(typ.Typ3(), bz)
+		_n, err = consumeAny(typ, bz)
 		if slide(&bz, &n, _n) && err != nil {
 			return
 		}
@@ -863,27 +832,20 @@ func consumeFieldKey(bz []byte) (typ Typ3, n int, err error) {
 
 //----------------------------------------
 
-func DecodeDisambPrefixBytes(bz []byte) (db DisambBytes, hasDb bool, pb PrefixBytes, typ Typ3, hasPb bool, isNil bool, n int, err error) {
-	// Special case: nil
-	if len(bz) >= 2 && bz[0] == 0x00 && bz[1] == 0x00 {
-		isNil = true
-		n = 2
-		return
-	}
+func DecodeDisambPrefixBytes(bz []byte) (db DisambBytes, hasDb bool, pb PrefixBytes, hasPb bool, n int, err error) {
 	// Validate
 	if len(bz) < 4 {
-		err = errors.New("EOF reading prefix bytes.")
+		err = errors.New("EOF while reading prefix bytes.")
 		return // hasPb = false
 	}
 	if bz[0] == 0x00 { // Disfix
 		// Validate
 		if len(bz) < 8 {
-			err = errors.New("EOF reading disamb bytes.")
+			err = errors.New("EOF while reading disamb bytes.")
 			return // hasPb = false
 		}
 		copy(db[0:3], bz[1:4])
 		copy(pb[0:4], bz[4:8])
-		pb, typ = pb.SplitTyp3()
 		hasDb = true
 		hasPb = true
 		n = 8
@@ -891,7 +853,6 @@ func DecodeDisambPrefixBytes(bz []byte) (db DisambBytes, hasDb bool, pb PrefixBy
 	} else { // Prefix
 		// General case with no disambiguation
 		copy(pb[0:4], bz[0:4])
-		pb, typ = pb.SplitTyp3()
 		hasDb = false
 		hasPb = true
 		n = 4
@@ -923,40 +884,9 @@ func decodeFieldNumberAndTyp3(bz []byte) (num uint32, typ Typ3, n int, err error
 	return
 }
 
-// Consume typ4 byte and error if it doesn't match rt.
-func decodeTyp4AndCheck(rt reflect.Type, bz []byte, opts FieldOptions) (ptr bool, n int, err error) {
-	var typ = Typ4(0x00)
-	typ, n, err = decodeTyp4(bz)
-	if err != nil {
-		return
-	}
-	var typWanted = typeToTyp4(rt, opts)
-	if typWanted != typ {
-		err = fmt.Errorf("unexpected Typ4. want: %v, got: %v", typWanted, typ)
-		return
-	}
-	ptr = (typ & 0x08) != 0
-	return
-}
-
-// Read Typ4 byte.
-func decodeTyp4(bz []byte) (typ Typ4, n int, err error) {
-	if len(bz) == 0 {
-		err = errors.New("EOF reading typ4 byte")
-		return
-	}
-	if bz[0]&0xF0 != 0 {
-		err = errors.New("invalid non-zero nibble reading typ4 byte")
-		return
-	}
-	typ = Typ4(bz[0])
-	n = 1
-	return
-}
-
 // Error if typ doesn't match rt.
-func checkTyp3(rt reflect.Type, typ Typ3, opts FieldOptions) (err error) {
-	typWanted := typeToTyp3(rt, opts)
+func checkTyp3(rt reflect.Type, typ Typ3, fopts FieldOptions) (err error) {
+	typWanted := typeToTyp3(rt, fopts)
 	if typ != typWanted {
 		err = fmt.Errorf("unexpected Typ3. want %v, got %v", typWanted, typ)
 	}
@@ -966,7 +896,7 @@ func checkTyp3(rt reflect.Type, typ Typ3, opts FieldOptions) (err error) {
 // Read typ3 byte.
 func decodeTyp3(bz []byte) (typ Typ3, n int, err error) {
 	if len(bz) == 0 {
-		err = fmt.Errorf("EOF reading typ3 byte")
+		err = fmt.Errorf("EOF while reading typ3 byte")
 		return
 	}
 	if bz[0]&0xF8 != 0 {
@@ -983,7 +913,7 @@ func decodeTyp3(bz []byte) (typ Typ3, n int, err error) {
 // other values will error.
 func decodeNumNilBytes(bz []byte) (numNil int64, n int, err error) {
 	if len(bz) == 0 {
-		err = errors.New("EOF reading nil byte(s)")
+		err = errors.New("EOF while reading nil byte(s)")
 		return
 	}
 	if bz[0] == 0x00 {